log gather call in case of exceptioin or reraise
