+++ services/web/server/src/simcore_service_webserver/resource_manager/garbage_collector.py
@@ -12,6 +12,7 @@ import logging
 from aiohttp import web
 
 from servicelib.observer import emit
+from servicelib.utils import logged_gather
 
 from .config import APP_GARBAGE_COLLECTOR_KEY, get_garbage_collector_interval
 from .registry import RedisResourceRegistry, get_registry
@@ -53,10 +54,8 @@ async def collect_garbage(registry: RedisResourceRegistry, app: web.Application)
                     logger.debug(
                         "removing resource entry: %s: %s", other_keys, resources
                     )
-                    results = await asyncio.gather(*remove_tasks, return_exceptions=True)
-                    for value in results:
-                        if isinstance(value, Exception):
-                            logger.error("Exception occured while remving resource: %s", value)
+                    logged_gather(*remove_tasks, reraise=False)
+
                 logger.debug(
                     "the resources %s:%s of %s may be now safely closed",
                     resource_name,
